requirements.txt
streamlit
google-genai
gTTS
pydub
streamlit-mic-recorder
pytest
pytest-mock
